// design-system/projects/kakal-ui/src/lib/form/models/form-events.ts
export enum FormActions {
  DEFAULT = 'default',
  CREATE = 'create',
  EDIT = 'edit',
  ADD = 'add',
  DELETE = 'delete',
  SUBMIT = 'submit',
  DISABLED = 'disable',
  CLEAR = 'clear',
  CLOSE = 'close',
  UPDATE = 'update',
  CHANGE = 'valueChange',
  SELECTED = 'optionSelected',
  MULTI_SELECTED = 'multiOptionSelected',
}